src/requirements.txt
pytz
python-dotenv
boto3
twilio
selenium
mailjet_rest